fix(T099): sync vendor directory with go.mod and update vendorHash
go.mod was updated with gomobile dependencies (x/mobile, x/image, x/mod,
x/tools, x/net upgrade) but vendor/ was not re-synced. This broke nix build
which uses vendored deps, cascading to all NixOS VM tests. Run go mod vendor
and update vendorHash in both nix/package.nix and nix/phonesim.nix.

Also add vendor/ to gitleaks allowlist since vendored third-party code
contains false positive secret detections (age library public keys).

+++ android/app/src/main/java/com/nixkey/bridge/GoPhoneServer.kt
@@ -50,7 +50,7 @@ class KeyStoreAdapter(
         return keyList
     }
 
-    override fun sign(fingerprint: String?, data: ByteArray?, flags: Int): ByteArray {
+    override fun sign(fingerprint: String?, data: ByteArray?, flags: Long): ByteArray {
         val fp = fingerprint ?: throw IllegalArgumentException("fingerprint is null")
         val d = data ?: throw IllegalArgumentException("data is null")
         val keys = keyManager.listKeys()
